+++ desktop/electron.vite.config.ts
@@ -6,7 +6,13 @@ import tailwindcss from "@tailwindcss/vite";
 
 export default defineConfig({
   main: {
-    plugins: [externalizeDepsPlugin()]
+    plugins: [externalizeDepsPlugin()],
+    resolve: {
+      alias: {
+        "@/lib": resolve("src/main/lib"),
+        "@shared": resolve("src/shared")
+      }
+    }
   },
   preload: {
     plugins: [externalizeDepsPlugin()]
@@ -14,7 +20,15 @@ export default defineConfig({
   renderer: {
     resolve: {
       alias: {
-        "@renderer": resolve("src/renderer/src")
+        "@renderer": resolve("src/renderer/src"),
+        "@shared": resolve("src/shared"),
+        "@/hooks": resolve("src/renderer/src/hooks"),
+        "@/assets": resolve("src/renderer/src/assets"),
+        "@/store": resolve("src/renderer/src/store"),
+        "@/components": resolve("src/renderer/src/components"),
+        "@/mocks": resolve("src/renderer/src/mocks"),
+        "@/lib": resolve("src/renderer/src/lib"),
+        "@/types": resolve("src/renderer/src/@types")
       }
     },
     plugins: [react(), tailwindcss()]