+++ desktop/src/preload/index.d.ts
@@ -1,8 +1,17 @@
-import { ElectronAPI } from '@electron-toolkit/preload'
+// import { electronAPI } from "@electron-toolkit/preload";
+import { Input, Output } from "@shared/types";
+
+export interface ApplicationInterface {
+  // electron: typeof electronAPI;
+
+  api: {
+    getMonacoBasePath: () => string;
+    executeCode: (payload: Input) => Promise<Output>;
+  };
+}
 
 declare global {
   interface Window {
-    electron: ElectronAPI
-    api: unknown
+    applicationApi: ApplicationInterface;
   }
 }